+++ src/classes/job.ts
@@ -1,4 +1,4 @@
-import { Redis, Pipeline } from 'ioredis';
+import { Pipeline } from 'ioredis';
 import { debuglog } from 'util';
 import { RetryErrors } from '../enums';
 import { BackoffOptions, JobsOptions, WorkerOptions } from '../interfaces';
@@ -8,6 +8,7 @@ import { QueueEvents } from './queue-events';
 import { Backoffs } from './backoffs';
 import { MinimalQueue, ParentOpts, Scripts } from './scripts';
 import { fromPairs } from 'lodash';
+import { RedisClient } from './redis-connection';
 
 const logger = debuglog('bull');
 
@@ -53,18 +54,48 @@ export interface MoveToChildrenOpts {
 }
 
 export class Job<T = any, R = any, N extends string = string> {
+  /**
+   * The progress a job has performed so far.
+   */
   progress: number | object = 0;
+
+  /**
+   * The value returned by the processor when processing this job.
+   */
   returnvalue: R = null;
+
+  /**
+   * Stacktrace for the error (for failed jobs).
+   */
   stacktrace: string[] = null;
+
+  /**
+   * Timestamp when the job was created (unless overrided with job options).
+   */
   timestamp: number;
 
+  /**
+   * Number of attempts after the job has failed.
+   */
   attemptsMade = 0;
+
+  /**
+   * Reason for failing.
+   */
   failedReason: string;
+
+  /**
+   * Timestamp for when the job finished (completed or failed).
+   */
   finishedOn?: number;
+
+  /**
+   * Timestamp for when the job was processed.
+   */
   processedOn?: number;
 
   /**
-   * Fully qualified key pointing to the parent of this job.
+   * Fully qualified key (including the queue prefix) pointing to the parent of this job.
    */
   parentKey?: string;
 
@@ -136,7 +167,12 @@ export class Job<T = any, R = any, N extends string = string> {
     const multi = client.multi();
 
     for (const job of jobInstances) {
-      job.addJob(<Redis>(multi as unknown));
+      job.addJob(<RedisClient>(multi as unknown), {
+        parentKey: job.parentKey,
+        parentDependenciesKey: job.parentKey
+          ? `${job.parentKey}:dependencies`
+          : '',
+      });
     }
 
     const result = (await multi.exec()) as [null | Error, string][];
@@ -451,7 +487,7 @@ export class Job<T = any, R = any, N extends string = string> {
   /**
    * Returns a promise the resolves when the job has finished. (completed or failed).
    */
-  async waitUntilFinished(queueEvents: QueueEvents, ttl?: number) {
+  async waitUntilFinished(queueEvents: QueueEvents, ttl?: number): Promise<R> {
     await this.queue.waitUntilReady();
 
     const jobId = this.id;
@@ -461,7 +497,9 @@ export class Job<T = any, R = any, N extends string = string> {
         timeout = setTimeout(
           () =>
             onFailed(
+              /* eslint-disable max-len */
               `Job wait ${this.name} timed out before finishing, no finish notification arrived after ${ttl}ms (id=${jobId})`,
+              /* eslint-enable max-len */
             ),
           ttl,
         );
@@ -585,7 +623,7 @@ export class Job<T = any, R = any, N extends string = string> {
     return Scripts.isJobInList(this.queue, this.queue.toKey(list), this.id);
   }
 
-  addJob(client: Redis, parentOpts?: ParentOpts): string {
+  addJob(client: RedisClient, parentOpts?: ParentOpts): string {
     const queue = this.queue;
 
     const jobData = this.asJSON();