+++ src/classes/worker.ts
@@ -1,11 +1,11 @@
 import * as fs from 'fs';
 import { Redis } from 'ioredis';
 import * as path from 'path';
-import { Processor, WorkerOptions } from '../interfaces';
+import { Processor, WorkerOptions, GetNextJobOptions } from '../interfaces';
 import { QueueBase, Repeat } from './';
 import { ChildPool } from './child-pool';
 import { Job, JobJsonRaw } from './job';
-import { RedisConnection } from './redis-connection';
+import { RedisConnection, RedisClient } from './redis-connection';
 import sandbox from './sandbox';
 import { Scripts } from './scripts';
 import { v4 } from 'uuid';
@@ -91,7 +91,7 @@ export class Worker<
     this.on('error', err => console.error(err));
   }
 
-  async waitUntilReady() {
+  async waitUntilReady(): Promise<RedisClient> {
     await super.waitUntilReady();
     return this.blockingConnection.client;
   }
@@ -119,7 +119,8 @@ export class Worker<
 
     // IDEA, How to store metadata associated to a worker.
     // create a key from the worker ID associated to the given name.
-    // We keep a hash table bull:myqueue:workers where every worker is a hash key workername:workerId with json holding
+    // We keep a hash table bull:myqueue:workers where
+    // every worker is a hash key workername:workerId with json holding
     // metadata of the worker. The worker key gets expired every 30 seconds or so, we renew the worker metadata.
     //
     try {
@@ -173,16 +174,20 @@ export class Worker<
    * @param token worker token to be assigned to retrieved job
    * @returns a Job or undefined if no job was available in the queue.
    */
-  async getNextJob(token: string) {
+  async getNextJob(token: string, { block = true }: GetNextJobOptions = {}) {
     if (this.paused) {
-      await this.paused;
+      if (block) {
+        await this.paused;
+      } else {
+        return;
+      }
     }
 
     if (this.closing) {
       return;
     }
 
-    if (this.drained) {
+    if (this.drained && block) {
       try {
         const jobId = await this.waitForJob();