+++ src/test/test_job.ts
@@ -19,11 +19,11 @@ const sinon = require('sinon');
 chai.use(chaiAsPromised);
 const expect = chai.expect;
 
-describe('Job', function() {
+describe('Job', function () {
   let queue: Queue;
   let queueName: string;
 
-  beforeEach(async function() {
+  beforeEach(async function () {
     queueName = 'test-' + v4();
     queue = new Queue(queueName);
   });
@@ -33,13 +33,13 @@ describe('Job', function() {
     await removeAllQueueData(new IORedis(), queueName);
   });
 
-  describe('.create', function() {
+  describe('.create', function () {
     const timestamp = 1234567890;
     let job: Job;
     let data: any;
     let opts: JobsOptions;
 
-    beforeEach(async function() {
+    beforeEach(async function () {
       data = { foo: 'bar' };
       opts = { timestamp };
 
@@ -47,7 +47,7 @@ describe('Job', function() {
       job = createdJob;
     });
 
-    it('saves the job in redis', async function() {
+    it('saves the job in redis', async function () {
       const storedJob = await Job.fromId(queue, job.id);
       expect(storedJob).to.have.property('id');
       expect(storedJob).to.have.property('data');
@@ -57,7 +57,7 @@ describe('Job', function() {
       expect(storedJob.opts.timestamp).to.be.equal(timestamp);
     });
 
-    it('should use the custom jobId if one is provided', async function() {
+    it('should use the custom jobId if one is provided', async function () {
       const customJobId = 'customjob';
       const createdJob = await Job.create(queue, 'test', data, {
         jobId: customJobId,
@@ -190,8 +190,8 @@ describe('Job', function() {
     });
   });
 
-  describe('.update', function() {
-    it('should allow updating job data', async function() {
+  describe('.update', function () {
+    it('should allow updating job data', async function () {
       const job = await Job.create<{ foo?: string; baz?: string }>(
         queue,
         'test',
@@ -204,8 +204,8 @@ describe('Job', function() {
     });
   });
 
-  describe('.remove', function() {
-    it('removes the job from redis', async function() {
+  describe('.remove', function () {
+    it('removes the job from redis', async function () {
       const job = await Job.create(queue, 'test', { foo: 'bar' });
       await job.remove();
       const storedJob = await Job.fromId(queue, job.id);
@@ -215,15 +215,15 @@ describe('Job', function() {
 
   // TODO: Add more remove tests
 
-  describe('.progress', function() {
-    it('can set and get progress as number', async function() {
+  describe('.progress', function () {
+    it('can set and get progress as number', async function () {
       const job = await Job.create(queue, 'test', { foo: 'bar' });
       await job.updateProgress(42);
       const storedJob = await Job.fromId(queue, job.id);
       expect(storedJob.progress).to.be.equal(42);
     });
 
-    it('can set and get progress as object', async function() {
+    it('can set and get progress as object', async function () {
       const job = await Job.create(queue, 'test', { foo: 'bar' });
       await job.updateProgress({ total: 120, completed: 40 });
       const storedJob = await Job.fromId(queue, job.id);
@@ -249,8 +249,8 @@ describe('Job', function() {
     });
   });
 
-  describe('.moveToCompleted', function() {
-    it('marks the job as completed and returns new job', async function() {
+  describe('.moveToCompleted', function () {
+    it('marks the job as completed and returns new job', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       await Job.create(queue, 'test', { foo: 'bar' });
@@ -272,7 +272,7 @@ describe('Job', function() {
      * Verify moveToFinished use default value for opts.maxLenEvents
      * if it does not exist in meta key (or entire meta key is missing).
      */
-    it('should not fail if queue meta key is missing', async function() {
+    it('should not fail if queue meta key is missing', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       await Job.create(queue, 'test', { color: 'red' });
@@ -291,21 +291,21 @@ describe('Job', function() {
         { idx: 1, baz: 'something' },
       ];
       const token = 'my-token';
-  
+
       const parentQueueName = 'parent-queue';
-      
+
       const parentQueue = new Queue(parentQueueName);
-  
+
       const parentWorker = new Worker(parentQueueName);
       const childrenWorker = new Worker(queueName);
       await parentWorker.waitUntilReady();
       await childrenWorker.waitUntilReady();
-  
+
       const data = { foo: 'bar' };
       const parent = await Job.create(parentQueue, 'testParent', data);
       const parentKey = getParentKey({
         id: parent.id,
-        queue: 'bull:' + parentQueueName
+        queue: 'bull:' + parentQueueName,
       });
       const client = await queue.client;
       const child1 = new Job(queue, 'testJob1', values[0]);
@@ -328,12 +328,14 @@ describe('Job', function() {
       const isActive = await job.isActive();
       expect(isActive).to.be.equal(true);
 
-      await expect(job.moveToCompleted('return value', token)).to.be.rejectedWith(`Job ${job.id} has pending dependencies finished`);
+      await expect(
+        job.moveToCompleted('return value', token),
+      ).to.be.rejectedWith(`Job ${job.id} has pending dependencies finished`);
 
       const isCompleted = await job.isCompleted();
 
       expect(isCompleted).to.be.false;
-  
+
       await childrenWorker.close();
       await parentWorker.close();
       await parentQueue.close();
@@ -341,8 +343,8 @@ describe('Job', function() {
     });
   });
 
-  describe('.moveToFailed', function() {
-    it('marks the job as failed', async function() {
+  describe('.moveToFailed', function () {
+    it('marks the job as failed', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       await Job.create(queue, 'test', { foo: 'bar' });
@@ -357,7 +359,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('moves the job to wait for retry if attempts are given', async function() {
+    it('moves the job to wait for retry if attempts are given', async function () {
       const queueEvents = new QueueEvents(queueName);
       await queueEvents.waitUntilReady();
 
@@ -388,7 +390,7 @@ describe('Job', function() {
       await queueEvents.close();
     });
 
-    it('marks the job as failed when attempts made equal to attempts given', async function() {
+    it('marks the job as failed when attempts made equal to attempts given', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       await Job.create(queue, 'test', { foo: 'bar' }, { attempts: 1 });
@@ -403,7 +405,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('moves the job to delayed for retry if attempts are given and backoff is non zero', async function() {
+    it('moves the job to delayed for retry if attempts are given and backoff is non zero', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       await Job.create(
@@ -425,7 +427,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('applies stacktrace limit on failure', async function() {
+    it('applies stacktrace limit on failure', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       const stackTraceLimit = 1;
@@ -446,7 +448,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('saves error stacktrace', async function() {
+    it('saves error stacktrace', async function () {
       const worker = new Worker(queueName);
       const token = 'my-token';
       await Job.create(queue, 'test', { foo: 'bar' });
@@ -478,7 +480,7 @@ describe('Job', function() {
       expect(isWaiting).to.be.equal(true);
     });
 
-    it('should process a promoted job according to its priority', async function() {
+    it('should process a promoted job according to its priority', async function () {
       const queueScheduler = new QueueScheduler(queueName);
       await queueScheduler.waitUntilReady();
 
@@ -528,7 +530,9 @@ describe('Job', function() {
       try {
         await job.promote();
         throw new Error('Job should not be promoted!');
-      } catch (err) {}
+      } catch (err) {
+        return;
+      }
     });
 
     it('should promote delayed job to the right queue if queue is paused', async () => {
@@ -730,19 +734,19 @@ describe('Job', function() {
   });
   */
 
-  describe('.finished', function() {
+  describe('.finished', function () {
     let queueEvents: QueueEvents;
 
-    beforeEach(async function() {
+    beforeEach(async function () {
       queueEvents = new QueueEvents(queueName);
       await queueEvents.waitUntilReady();
     });
 
-    afterEach(async function() {
+    afterEach(async function () {
       await queueEvents.close();
     });
 
-    it('should resolve when the job has been completed', async function() {
+    it('should resolve when the job has been completed', async function () {
       const worker = new Worker(queueName, async job => 'qux');
 
       const job = await queue.add('test', { foo: 'bar' });
@@ -754,7 +758,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('should resolve when the job has been completed and return object', async function() {
+    it('should resolve when the job has been completed and return object', async function () {
       const worker = new Worker(queueName, async job => ({ resultFoo: 'bar' }));
 
       const job = await queue.add('test', { foo: 'bar' });
@@ -767,7 +771,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('should resolve when the job has been delayed and completed and return object', async function() {
+    it('should resolve when the job has been delayed and completed and return object', async function () {
       const worker = new Worker(queueName, async job => {
         await delay(300);
         return { resultFoo: 'bar' };
@@ -783,7 +787,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('should resolve when the job has been completed and return string', async function() {
+    it('should resolve when the job has been completed and return string', async function () {
       const worker = new Worker(queueName, async job => 'a string');
 
       const job = await queue.add('test', { foo: 'bar' });
@@ -796,7 +800,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('should reject when the job has been failed', async function() {
+    it('should reject when the job has been failed', async function () {
       const worker = new Worker(queueName, async job => {
         await delay(500);
         throw new Error('test error');
@@ -814,7 +818,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('should resolve directly if already processed', async function() {
+    it('should resolve directly if already processed', async function () {
       const worker = new Worker(queueName, async job => ({ resultFoo: 'bar' }));
 
       const job = await queue.add('test', { foo: 'bar' });
@@ -828,7 +832,7 @@ describe('Job', function() {
       await worker.close();
     });
 
-    it('should reject directly if already processed', async function() {
+    it('should reject directly if already processed', async function () {
       const worker = new Worker(queueName, async job => {
         throw new Error('test error');
       });